Change value decimal field to integer field.

+++ trade/models.py
@@ -12,9 +12,7 @@ class Good(models.Model):
     description = models.TextField()
     modified = models.DateField(
         auto_now=True)
-    value = models.DecimalField(
-        max_digits=10,
-        decimal_places=2)
+    value = models.IntegerField()
 
 class UserInfo(models.Model):
     user = models.ForeignKey(User)
@@ -26,9 +24,7 @@ class Need(models.Model):
     detail = models.TextField()
     modified = models.DateField(
         auto_now=True)
-    value = models.DecimalField(
-        max_digits=10,
-        decimal_places=2)
+    value = models.IntegerField()
 
 class Transaction(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
